requirements.txt
# FastAPI and core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database (using psycopg3 instead of psycopg2)
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0

# Environment variables
python-dotenv==1.0.1

# HTTP requests
httpx==0.27.2
aiohttp==3.10.10

# Image handling
cloudinary==1.41.0
Pillow==11.0.0

# SMS (Africa's Talking)
africastalking==1.2.6

# Payment gateway
requests==2.32.3

# Validation
pydantic==2.10.0
pydantic-settings==2.6.1
email-validator==2.2.0

# WebSockets
websockets==14.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# handling images 
cloudinary

#handling payments
paynow
